backend/requirements.txt
fastapi
uvicorn
requests
celery
traveltimepy
pandas
mysql-connector-python
coverage
pytest
